Renaming Plot.py -> PlotWidget.py
